[Changed] - Rendering method to GL_TRIANGLES [Changed] - Terrain rendering code now (temporarily) uses unsigned ints [Changed] - Camera spawn height
Temporarily disabled the normal calculations. Will be re-enabled later.

+++ LAGGED/src/Platform/OpenGL/Renderer/GL_Surface.h
@@ -37,7 +37,7 @@ namespace LAG
 
 		std::vector<VertexData> m_VertexData;
 		//std::vector<float> m_Normals;
-		std::vector<unsigned short> m_Indices;
+		std::vector<unsigned int> m_Indices;
 
 		unsigned int m_VBO = 0;
 		unsigned int m_EBO = 0;
